Взаимодействие с объектами через lua, читай описание
Обновил холодильник и раковину в unsafeflat
Добавил удержание backspace в полях ввода

По поводу взаимодействия с объектами:
Для изменения значения пишем objVar("имя объекта", "имя переменной", а тут значение)
Чтобы получить значение переменной тоже самое, но без самого последнего аргумента
Если с последним аргументом, то значение тоже вернётся, просто  это будет третий агрумент по сути

Существуют такие переменные объектов
Для всех
pos {num, num, num}
x num
y num
z num
activable bol
activateRadius num
clickable bol
pointable bol
animateWhenPaused bol

Для физических
hp num
speed {num, num, num}
rotY num
radius num
height num
physics bol
pushable bol
canPush bol

+++ src/code/game/world/entities/PhysEntity.java
@@ -13,7 +13,7 @@ import code.utils.FPS;
  */
 public class PhysEntity extends Entity {
     
-    public int hp = 100;
+    public float hp = 100;
     public Vector3D speed = new Vector3D();
     public float rotY;
     public float radius, height;
